+++ GraphicsCVarControl/Source/GraphicsCVarControlEditor/Private/GraphicsCVarReportExporter.h
@@ -1,0 +1,21 @@
+#pragma once
+
+#include "CoreMinimal.h"
+
+class FGraphicsCVarProfiler;
+
+struct FGraphicsCVarReportExportResult
+{
+	bool bSuccess = false;
+	FString MarkdownPath;
+	FString JsonPath;
+	FString ErrorMessage;
+};
+
+class FGraphicsCVarReportExporter
+{
+public:
+	static FGraphicsCVarReportExportResult ExportComparisonReport(
+		const FGraphicsCVarProfiler& Profiler,
+		double HighlightThresholdMs);
+};